Use a node executable instead of a bash one so that bin symlinks work
This resolves a problem running symlinks to bin/libyear-npm. E.g. if the package
was installed globally with `npm install -g`, the symlink at
`/usr/local/bin/libyear-npm` would result in an error like:

> Error: Cannot find module '/usr/local/lib/libyear-npm.js'

`require` seems to "just work" when it comes to resolving symlinks and
module paths, and the node executable is even simpler than the bash one.