Allow calculator to be used programmatically
This commit changes calculator to return an object with the result. This allows for programmatic use, i.e. using this package in your application and doing with the result whatever you want. For this, I built up an object with the results. But because we want to keep seeing the output as it is calculated, I introduced an event emitter to which the main function subscribes. It can then write to stdout as the results are coming in.

I've also chosen to throw errors and handle everything in the main function (instead of using stdout and exiting). In case someone uses this as a library in their application, it's up to them to handle errors.